requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
pydantic
python-dotenv
email-validator
reportlab
weasyprint
aiofiles
Jinja2
requests
